Use weighted treverhines/RBF instead of scipy rbf
this is all pretty stab in the dark, but gaussian instead of linear
makes sense to me, and the confidence score to stdev mapping is chosen
to make some kind of sense in log space. But in large part it was just
chosen based on what makes a good-looking map.

There's also a big-ish Dockerfile refactor hidden in this branch: we
only run "conda create" once, the naturalearth data download is moved
further up to reduce work, and we now need build-essential to build RBF.