Apply regularization to the global scaling WLS
This represents the idea that our data is sparse and noisy and we should
give some credence to the idea that the IRI model is basically okay. It
also reduces the amount by which a large change in one point (or a
station coming online / aging out) can change the values all around the
globe, so it should reduce the "jitter" from one run to the next.